implement solid color rendering for swaybg
